+++ test/e2e/navigator/scenarios.js
@@ -63,5 +63,23 @@
         expect(page3.isPresent()).not.toBeTruthy();
       });
     });
+
+    it('should emit events', function() {
+      browser.get(path);
+
+      var pops = element(by.id('pops')),
+        pushes = element(by.id('pushes'));
+
+      expect(pops.getText()).toBe('0');
+      expect(pushes.getText()).toBe('1');
+
+      element(by.id('btn1')).click();
+      browser.wait(EC.visibilityOf(element(by.id('btn2'))));
+      element(by.id('btn2')).click();
+      browser.wait(EC.textToBePresentInElement(pops, '1'));
+
+      expect(pops.getText()).toBe('1');
+      expect(pushes.getText()).toBe('2');
+    });
   });
 })();